requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-socketio>=5.11.0
motor>=3.3.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
httpx>=0.26.0
openai>=1.12.0
anthropic>=0.18.0
